feat(brain): preference-weighted voting for investor personas by asset class
Replaces flat 1-vote-per-agent Panel B counting with asset-class-aware weights.

_PERSONA_WEIGHTS table (stock / crypto):
- Buffett/Munger: stock=1.5 crypto=0.1  (stock legends; "rat poison" on crypto)
- Lynch:          stock=1.5 crypto=0.5  (GARP stocks; crypto outside his domain)
- Ackman:         stock=1.2 crypto=0.3  (activist stocks; minimal crypto conviction)
- Cohen:          stock=1.0 crypto=1.2  (pure momentum — works on any liquid asset)
- Dalio:          stock=1.2 crypto=0.8  (all-weather macro; small crypto as hedge)
- Wood:           stock=1.0 crypto=1.8  (strongest crypto/innovation conviction)
- Bogle:          stock=1.0 crypto=0.0  (passive indexer; EXCLUDED from crypto)
- Soros:          stock=1.0 crypto=1.5  (reflexivity ideal for crypto boom/bust)
- Druckenmiller:  stock=1.0 crypto=1.5  (concentrated macro; publicly owns BTC)
- Simons:         stock=1.0 crypto=1.3  (quant; statistical patterns ideal for crypto)
- Templeton:      stock=1.2 crypto=0.5  (contrarian value; stocks his domain)

Key changes:
- _count_votes_weighted(): Panel B uses float weighted tallies per asset class
- _count_votes(): Panel A unchanged — always 1.0 (domain-agnostic)
- _total_b_weight() / _total_system_weight(): compute effective pool per asset class
- _aggregate_dual_panel(): accepts asset_class; proportional b_abstaining threshold
  (83.3% = 10/12 ratio maintained: stock=11.75 crypto=8.0)
- _paper_risk_manager(): confidence = votes / _total_system_weight() (not /27.0)
- signal.py: votes_for_action int→float (weighted tally is float)

Emergent behaviour (verified):
- Buffett+Munger+Bogle NEUTRAL on crypto → 0.2 weighted units → no b_abstaining ✓
- Wood+Soros+Druckenmiller+Simons+Cohen all BULLISH on crypto → 7.3 + Panel A ✓
- Crypto HOT threshold requires 44.7% agreement (more conservative than stock 37.8%) ✓
- Bogle on crypto = 0.0 (zero) — excluded entirely, can't block a crypto trade ✓

+++ brain/signal.py
@@ -18,9 +18,10 @@ class TradingSignal:
     # HITL tier — driven by vote count + deterministic regime, NOT LLM confidence
     tier: Literal["HOT", "WARM", "COLD"] = "WARM"
 
-    # Vote tally — Panel A (11 analysts) + Panel B (8 investor personas) = combined 19
+    # Vote tally — Panel A (15 analysts × 1.0) + Panel B (12 personas × preference weight)
+    # combined tally values are float due to weighted Panel B votes
     vote_tally: dict = field(default_factory=dict)       # combined {bullish: N, bearish: N, neutral: N}
-    votes_for_action: int = 0                             # combined count of agents that agree with action
+    votes_for_action: float = 0.0                        # weighted vote count agreeing with action
     regime_label: str = "UNKNOWN"                         # TRENDING_UP / TRENDING_DOWN / RANGING / HIGH_VOLATILITY
 
     # Dual-panel fields
